+++ packages/cloud-cognitive/src/components/AddSelect/AddSelectList.js
@@ -19,25 +19,29 @@ import {
 import { ChevronRight16, View16 } from '@carbon/icons-react';
 import PropTypes from 'prop-types';
 import cx from 'classnames';
-import { UserProfileImage } from '../UserProfileImage';
 import { pkg } from '../../settings';
+import { UserProfileImage } from '../UserProfileImage';
+
+const blockClass = `${pkg.prefix}--add-select__selections`;
 const componentName = 'AddSelectList';
 
 export let AddSelectList = ({
+  appliedModifiers,
   filteredItems,
   metaIconDescription,
   modifiers,
   multi,
   multiSelection,
   navIconDescription,
-  path,
+  parentId,
+  setAppliedModifiers,
   setDisplayMetaPanel,
   setMultiSelection,
-  setPath,
+  setParentSelected,
   setSingleSelection,
   singleSelection,
 }) => {
-  const blockClass = `${pkg.prefix}--add-select__selections`;
+  const hasModifiers = modifiers?.options?.length > 0;
 
   const handleSingleSelection = (value) => {
     setSingleSelection(value);
@@ -53,33 +57,8 @@ export let AddSelectList = ({
     }
   };
 
-  const onNavigateItem = ({ id, title, parent }) => {
-    // if multi select
-    if (multi) {
-      // if top level reset the path
-      if (!parent) {
-        setPath([{ id, title }]);
-      } else {
-        const pathIds = path.map((p) => p.id);
-        // if item is already selected somewhere go back to that item
-        if (pathIds.includes(id)) {
-          const pathIdx = pathIds.findIndex((pathId) => pathId === id);
-          const newPath = [...path].splice(0, pathIdx + 1);
-          setPath([...newPath]);
-        } else {
-          // if the item is on the same level as another selected item start from the parent level
-          if (path.find((p) => p.parent === parent)) {
-            const parentIdx = path.findIndex((p) => p.id === parent);
-            const newPath = [...path].splice(0, parentIdx + 1);
-            setPath([...newPath, { id, title, parent }]);
-          } else {
-            setPath([...path, { id, title, parent }]);
-          }
-        }
-      }
-    } else {
-      setPath([...path, { id, title }]);
-    }
+  const onNavigateItem = ({ id, title }) => {
+    setParentSelected(id, title, parentId);
   };
 
   const isSelected = (id) => multiSelection.includes(id);
@@ -95,6 +74,16 @@ export let AddSelectList = ({
   });
 
   const getItemIcon = ({ icon: Icon }) => <Icon />;
+
+  const modifierHandler = (id, selectedItem) => {
+    const modifiersClone = [...appliedModifiers];
+    const modifierIdx = modifiersClone.findIndex((item) => item.id === id);
+    modifiersClone[modifierIdx] = {
+      id,
+      [modifiers.id]: selectedItem,
+    };
+    setAppliedModifiers(modifiersClone);
+  };
 
   return (
     <div className={`${blockClass}-wrapper`}>
@@ -144,15 +133,19 @@ export let AddSelectList = ({
                           </div>
                         </div>
                       </div>
-                      {modifiers?.options?.length && (
+                      {hasModifiers && (
                         <Dropdown
                           id={`${item.id}-modifier`}
                           type="inline"
-                          items={modifiers?.options}
+                          items={modifiers.options}
                           light
-                          label={modifiers?.label}
+                          label={modifiers.label}
                           disabled={!isSelected(item.id)}
                           className={`${blockClass}-dropdown ${blockClass}-hidden-hover`}
+                          initialSelectedItem={item[modifiers.id]}
+                          onChange={({ selectedItem }) =>
+                            modifierHandler(item.id, selectedItem)
+                          }
                         />
                       )}
                     </>
@@ -163,12 +156,13 @@ export let AddSelectList = ({
                       id={item.id}
                       value={item.value}
                       labelText={item.title}
-                      onChange={handleSingleSelection}
-                      selected={item.value === singleSelection}
+                      onChange={() => handleSingleSelection(item.id)}
+                      selected={item.id === singleSelection}
                     />
                   )}
                   {item.children && (
                     <Button
+                      className={`${blockClass}-view-children`}
                       renderIcon={ChevronRight16}
                       iconDescription={navIconDescription}
                       tooltipPosition="left"
@@ -182,6 +176,7 @@ export let AddSelectList = ({
                   {item.meta && (
                     <div className={`${blockClass}-hidden-hover`}>
                       <Button
+                        className={`${blockClass}-view-meta`}
                         renderIcon={View16}
                         iconDescription={metaIconDescription}
                         tooltipPosition="left"
@@ -204,16 +199,18 @@ export let AddSelectList = ({
 };
 
 AddSelectList.propTypes = {
+  appliedModifiers: PropTypes.array,
   filteredItems: PropTypes.array,
   metaIconDescription: PropTypes.string,
   modifiers: PropTypes.object,
   multi: PropTypes.bool,
   multiSelection: PropTypes.array,
   navIconDescription: PropTypes.string,
-  path: PropTypes.array,
+  parentId: PropTypes.string,
+  setAppliedModifiers: PropTypes.func,
   setDisplayMetaPanel: PropTypes.func,
   setMultiSelection: PropTypes.func,
-  setPath: PropTypes.func,
+  setParentSelected: PropTypes.func,
   setSingleSelection: PropTypes.func,
   singleSelection: PropTypes.string,
 };