+++ packages/cloud-cognitive/src/components/MultiAddSelect/MultiAddSelect.stories.js
@@ -93,7 +93,6 @@ const defaultProps = {
   onCloseButtonText: 'Cancel',
   onSubmit: (selections) => console.log(selections),
   onSubmitButtonText: 'Add',
-  removeIconDescription: 'Remove',
   searchResultsLabel: 'Search results',
   title: 'Add business terms',
 };
@@ -155,6 +154,18 @@ export const WithHierarchy = prepareStory(Template, {
                 size: '100',
                 icon: Document16,
                 tag: 'business',
+                children: {
+                  entries: [
+                    {
+                      id: '9000',
+                      value: '9000.html',
+                      title: '9000.html',
+                      fileType: 'html',
+                      size: '9000',
+                      icon: Document16,
+                    },
+                  ],
+                },
               },
               {
                 id: '1-2',
@@ -179,6 +190,18 @@ export const WithHierarchy = prepareStory(Template, {
           id: '2',
           value: 'folder 2',
           title: 'folder 2',
+          children: {
+            entries: [
+              {
+                id: '7000',
+                value: '7000.html',
+                title: '7000.html',
+                fileType: 'html',
+                size: '7000',
+                icon: Document16,
+              },
+            ],
+          },
         },
       ],
     },
@@ -190,6 +213,7 @@ export const WithModifiers = prepareStory(Template, {
     ...defaultProps,
     items: {
       modifiers: {
+        id: 'role',
         label: 'Role',
         options: ['editor', 'viewer', 'admin'],
       },
@@ -199,17 +223,20 @@ export const WithModifiers = prepareStory(Template, {
           value: '1',
           title: 'item 1',
           subtitle: 'item 1 subtitle',
+          role: 'editor',
         },
         {
           id: '2',
           value: '2',
           title: 'item 2',
+          role: 'editor',
         },
         {
           id: '3',
           value: '3',
           title: 'item 3',
           subtitle: 'item 3 subtitle',
+          role: 'admin',
         },
       ],
     },