+++ packages/cloud-cognitive/CHANGELOG.md
@@ -3,6 +3,110 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.23.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.23.0...@carbon/ibm-products@1.23.1) (2022-07-19)
+
+
+### Bug Fixes
+
+* update to Carbon v10 compatible versions to latest ([#2101](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2101)) ([feebce3](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/feebce38ac474917e29d201d9198e3baf9452a97))
+* **web-terminal:** spans 100% when less than 640px ([#2087](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2087)) ([c012d30](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/c012d306c0df936ef4916e5b31e7dd8517ebb3ca))
+* **WebTerminal:** small fixes for release review ([#2106](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2106)) ([6eb3346](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/6eb3346efaae41a1c38972620c4f9222419cce36))
+
+
+
+
+
+# [1.23.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.22.0...@carbon/ibm-products@1.23.0) (2022-07-12)
+
+
+### Bug Fixes
+
+* add select refactor ([#2068](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2068)) ([b7aa521](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/b7aa5217902ae4ca29c998692f8d7b2823c90f69))
+
+
+### Features
+
+* add multiline attribute to tagset ([#2079](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2079)) ([6c3286a](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/6c3286ae43f8e4996badf20dc4582b8133a36d0f))
+* align expandable rows ([#2039](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2039)) ([853176f](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/853176f951d13ab867bee63cb2b2a562fd7ce7a7))
+* index cloud cognitive ([#2063](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2063)) ([a743a20](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/a743a206a016685f3a8a2644bbdce054ad84ead1))
+* merged left panel content and datagrid on same horizontal level ([#2064](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2064)) ([356b101](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/356b10107055a8f72ee68f07851357405306b99e))
+
+
+
+
+
+# [1.22.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.21.0...@carbon/ibm-products@1.22.0) (2022-07-05)
+
+
+### Features
+
+* added center aligned columns component and centered new columns ([#2056](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2056)) ([9178df2](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/9178df24c2f54370ab4064ab659c77a9ac459377))
+
+
+
+
+
+# [1.21.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.20.3...@carbon/ibm-products@1.21.0) (2022-06-28)
+
+
+### Features
+
+* **WebTerminal:** Changes width to 640px ([#2066](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2066)) ([b741316](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/b741316b74dd8be0af2f4cfc61ad28c0a672e46b))
+
+
+
+
+
+## [1.20.3](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.20.2...@carbon/ibm-products@1.20.3) (2022-06-21)
+
+**Note:** Version bump only for package @carbon/ibm-products
+
+
+
+
+
+## [1.20.2](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.20.1...@carbon/ibm-products@1.20.2) (2022-06-15)
+
+
+### Bug Fixes
+
+* **Datagrid:** use checkbox in place of inline checkbox ([#2037](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2037)) ([1859ccf](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/1859ccf5f1f7181ed83aad0125268042fb5a7ca1))
+
+
+
+
+
+## [1.20.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.20.0...@carbon/ibm-products@1.20.1) (2022-06-14)
+
+
+### Bug Fixes
+
+* **Datagrid:** export datagrid hooks ([#2036](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2036)) ([44be0e1](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/44be0e1e31cf5061296238026901280bdd49bc75))
+
+
+
+
+
+# [1.20.0](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.19.1...@carbon/ibm-products@1.20.0) (2022-06-13)
+
+
+### Bug Fixes
+
+* **ComboButton:** remove getInstanceId and use our internal uuid util ([#2028](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2028)) ([c482bbd](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/c482bbd1378ad4b774a118e677e05ccad88c4675))
+* **CreateSidePanel:** use actions prop instead of ActionSet directly ([#2033](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2033)) ([8dbeabd](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/8dbeabdc3a2829e9a60207c9a36da58d9817add0))
+* header row horizontal scroll and scrollbar style ([#2032](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2032)) ([f6ea8d8](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/f6ea8d88ca11cc959d1c36c5d309f23414cc03e3))
+* hooked up add select modifier functionality ([#2019](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2019)) ([2cbda09](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/2cbda091755b6c1dde20e892c5b8568c0ab18e11))
+* **SidePanel:** check if window exists for reduced motion logic ([#2029](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2029)) ([6bc2b29](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/6bc2b29a3dc35ee7952cf8683e5f8a9e46e59e2c))
+
+
+### Features
+
+* inline edit release ([#2025](https://github.com/carbon-design-system/ibm-cloud-cognitive/issues/2025)) ([028a58f](https://github.com/carbon-design-system/ibm-cloud-cognitive/commit/028a58fc10fab8cd2f7e8846d0ef7611e15862ca))
+
+
+
+
+
 ## [1.19.1](https://github.com/carbon-design-system/ibm-cloud-cognitive/compare/@carbon/ibm-products@1.19.0...@carbon/ibm-products@1.19.1) (2022-06-06)